+++ CHANGELOG.md
@@ -6,7 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-06-23
+
 ### Added
 
--  Initial project structure
+-  Initial support for automation of checklist steps
+-  Perform validation of loaded map layers
+-  Generate report in JSON format
+-  Load checklists from remote server
+-  Default server with checklists for GeoCRIS
 -  Paver file for development
+-  Initial project structure
+
+
+[unreleased]: https://github.com/kartoza/qgis_checklist_checker/compare/v0.2.0...master
+[0.2.0]: https://github.com/kartoza/qgis_checklist_checker/-/tags/v0.2.0